icare_chatbot_mvp/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.7.1
python-dotenv==1.0.1
